perf: Move HLS elements structs from types.go to hlselement.go
