+++ decode.go
@@ -39,8 +39,8 @@ func ParsePlaylist(src Source) (*Playlist, error) {
 	playlist := &Playlist{
 		DoublyLinkedList: new(internal.DoublyLinkedList),
 		CurrentDateRange: new(internal.Node),
-		CurrentSegment:   new(Segment),
-		CurrentStreamInf: new(StreamInf),
+		CurrentSegment:   new(internal.Segment),
+		CurrentStreamInf: new(internal.StreamInf),
 		ProgramDateTime:  *new(time.Time),
 		MediaSequence:    0,
 		SegmentsCounter:  0,
