+++ types.go
@@ -1,9 +1,5 @@
 package go_m3u8
 
-import (
-	"time"
-)
-
 const (
 	m3u8IdentifierTag     = "#EXTM3U"
 	versionTag            = "#EXT-X-VERSION"
@@ -20,30 +16,3 @@ const (
 	cueInTag              = "#EXT-X-CUE-IN"
 	scteOutAttribute      = "SCTE35-OUT"
 )
-
-type StreamInf struct {
-	Codecs           []string
-	Bandwidth        string
-	AverageBandwidth string
-	Resolution       string
-	FrameRate        string
-	URI              string
-}
-
-type Segment struct {
-	Duration        float64
-	ProgramDateTime time.Time
-	MediaSequence   int
-	URI             string
-}
-
-type DateRange struct {
-	ID              string
-	Class           string
-	StartDate       time.Time
-	EndDate         time.Time
-	PlannedDuration float64
-	Scte35Out       string
-	Scte35In        string
-	MediaSequence   int
-}
